+++ example/single_file_example.py
@@ -9,9 +9,9 @@ from panther.openapi import OutputSchema
 from panther.openapi.urls import urls as openapi_urls
 from panther.request import Request
 from panther.response import Response
-from panther.throttling import Throttling
+from panther.throttling import Throttle
 
-InfoThrottling = Throttling(rate=5, duration=timedelta(minutes=1))
+InfoThrottle = Throttle(rate=5, duration=timedelta(minutes=1))
 
 
 @API()
@@ -31,7 +31,7 @@ class UserSerializer(BaseModel):
     name: str = "hi"
 
 
-@API(cache=True, throttling=InfoThrottling, input_model=UserSerializer, methods=['GET', 'POST', 'delete'])
+@API(throttling=InfoThrottle, input_model=UserSerializer, methods=['GET', 'POST', 'delete'])
 async def info(request: Request):
     """Hi from info"""
     data = {