+++ tests/test_throttling.py
@@ -1,11 +1,12 @@
 import asyncio
-from datetime import timedelta
+from datetime import timedelta, datetime
 from unittest import IsolatedAsyncioTestCase
 
 from panther import Panther
 from panther.app import API
 from panther.test import APIClient
-from panther.throttling import Throttling
+from panther.throttling import Throttle
+from panther.utils import round_datetime
 
 
 @API()
@@ -13,14 +14,20 @@ async def without_throttling_api():
     return 'ok'
 
 
-@API(throttling=Throttling(rate=3, duration=timedelta(seconds=3)))
+@API(throttling=Throttle(rate=3, duration=timedelta(seconds=1)))
 async def with_throttling_api():
+    return 'ok'
+
+
+@API(throttling=Throttle(rate=1, duration=timedelta(seconds=1)))
+async def throttling_headers_api():
     return 'ok'
 
 
 urls = {
     'without-throttling': without_throttling_api,
     'with-throttling': with_throttling_api,
+    'throttling-headers': throttling_headers_api,
 }
 
 
@@ -56,7 +63,7 @@ class TestThrottling(IsolatedAsyncioTestCase):
         res5 = await self.client.get('with-throttling')
         assert res5.status_code == 429
 
-        await asyncio.sleep(3)  # Sleep and try again
+        await asyncio.sleep(1)  # Sleep and try again
 
         res6 = await self.client.get('with-throttling')
         assert res6.status_code == 200
@@ -72,3 +79,17 @@ class TestThrottling(IsolatedAsyncioTestCase):
 
         res10 = await self.client.get('with-throttling')
         assert res10.status_code == 429
+
+    async def test_throttling_header(self):
+        await self.client.get('throttling-headers')
+
+        res = await self.client.get('throttling-headers')
+        assert res.status_code == 429
+        reset_time = round_datetime(datetime.now(), timedelta(seconds=1)) + timedelta(seconds=1)
+        assert res.headers == {
+            'Content-Type': 'application/json',
+            'Content-Length': '29',
+            'Access-Control-Allow-Origin': '*',
+            'Retry-After': str(int((reset_time - datetime.now()).total_seconds())),
+            'X-RateLimit-Reset': str(int(reset_time.timestamp()))
+        }